small fixes are review tool critizism
HU-Trace: show all connected records https://github.com/metasfresh/metasfresh-webui-api/issues/632
